refactor: reorganize services into core modules
