+++ packages/skaff-lib/src/actions/git/diff-project-from-template.ts
@@ -1,6 +1,6 @@
 import { ParsedFile, Result } from "../../lib";
 import { Project } from "../../models";
-import { diffProjectFromItsTemplate } from "../../services/project-diff-service";
+import { diffProjectFromItsTemplate } from "../../core/diffing/project-diff-service";
 
 export async function diffProjectFromTemplate(
   project: Project
