+++ packages/skaff-lib/src/actions/instantiate/prepare-update-diff.ts
@@ -1,6 +1,6 @@
 import { NewTemplateDiffResult, Result } from "../../lib";
 import { Project } from "../../models";
-import { generateUpdateTemplateDiff } from "../../services/project-diff-service";
+import { generateUpdateTemplateDiff } from "../../core/diffing/project-diff-service";
 
 export async function prepareUpdateDiff(
   project: Project,
