+++ packages/skaff-lib/src/core/generation/ProjectSettingsSynchronizer.ts
@@ -12,9 +12,9 @@ import {
   removeTemplateFromSettings,
   writeNewProjectSettings,
   writeNewTemplateToSettings,
-} from "../../services/project-settings-service";
+} from "../projects/project-settings-service";
 import { getLatestTemplateMigrationUuid } from "../templates/TemplateMigration";
-import { GeneratorOptions } from "../../services/template-generator-service";
+import { GeneratorOptions } from "./template-generator-service";
 
 export class ProjectSettingsSynchronizer {
   constructor(
