+++ packages/skaff-lib/src/core/infra/file-rollback-service.ts
@@ -1,0 +1,1 @@
+export { FileRollbackManager } from "../shared/FileRollbackManager";
