+++ packages/skaff-lib/src/core/infra/shell-service.ts
@@ -1,7 +1,7 @@
 import { exec } from "node:child_process";
 import { promisify } from "node:util";
-import { Result } from "../lib/types";
-import { logError } from "../lib/utils";
+import { Result } from "../../lib/types";
+import { logError } from "../../lib/utils";
 
 const asyncExec = promisify(exec);
