+++ packages/skaff-lib/src/index.ts
@@ -4,6 +4,6 @@ export * from "./lib";
 export * from "./actions";
 
 export { findTemplate, projectSearchPathKey } from "./utils/shared-utils";
-export { getCacheDirPath, getCacheDir, pathInCache, saveToCache } from "./services/cache-service";
-export { getRemoteCommitHash } from "./services/git-service";
-export type { CacheKey } from "./services/cache-service";
+export { getCacheDirPath, getCacheDir, pathInCache, saveToCache } from "./core/infra/cache-service";
+export { getRemoteCommitHash } from "./core/infra/git-service";
+export type { CacheKey } from "./core/infra/cache-service";
