--- packages/skaff-lib/src/services/file-rollback-service.ts
@@ -1,1 +1,0 @@
-export { FileRollbackManager } from "../core/shared/FileRollbackManager";
